add company add and company change.

+++ main-ms/src/main/java/com/vidnichuk/isogj/api/dto/model/UserCompanyDto.java
@@ -2,15 +2,16 @@ package com.vidnichuk.isogj.api.dto.model;
 
 
 public class UserCompanyDto {
-    private int idUserCompany;
+    private Long idUserCompany;
     private CompanyDto company;
     private PositionDto position;
 
-    public int getIdUserCompany() {
+
+    public Long getIdUserCompany() {
         return idUserCompany;
     }
 
-    public void setIdUserCompany(int idUserCompany) {
+    public void setIdUserCompany(Long idUserCompany) {
         this.idUserCompany = idUserCompany;
     }
